requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dm-tree==0.1.8
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.13.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.0.5
kiwisolver==1.4.8
libclang==16.0.6
llvmlite==0.44.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.15.0
packaging==23.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.3
py-spy==0.4.0
py4j==0.10.9.9
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==309
requests==2.31.0
rich==13.7.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==69.1.1
six==1.16.0
slicer==0.0.8
tensorboard-data-server==0.7.2
tensorflow==2.16.0rc0
tensorflow-intel==2.16.0rc0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.10.0
tzdata==2025.1
urllib3==2.2.1
Werkzeug==3.1.3
wget==3.2
wheel==0.42.0
wrapt==1.16.0
xgboost==2.1.4
